refactor(xlsx): extract number/date formatting from renderer.ts
Move cell-value and number/date format-code handling (~515 lines) into
src/number-format.ts: cellValueText, format-code tokenizer/applier,
built-in date formats, Japanese era/weekday rendering, Excel serial-date
conversion. Exposes formatCellValue; imports todaySerial/nowSerial from
formula.js.

renderer.ts drops from 4118 to 3603 lines. Typecheck passes, VRT
regression 65/65 at 0.0% diff.
